+++ components/index.ts
@@ -1,2 +1,2 @@
-export { MenuBar } from "./MenuBar";
+export { MenuBar } from "./Layout";
 export { Icon } from "./Icon";
